don't backup symlinks

+++ install.sh
@@ -3,19 +3,25 @@ function link_file {
     source="${PWD}/$1"
     target="${HOME}/${1/_/.}"
 
-    if [ -e "${target}" ]; then
+	# backup if the file is already there (and not a symlink)
+    if [ -e "${target}" -a ! -L "${target}" ]; then
+		echo "    backing up $target"
         mv $target $target.bak
     fi
 
+	echo "    installing $source to $target"
     ln -sf ${source} ${target}
 }
 
+echo "installing dotfile(s)..."
 for i in _$1*
 do
 	link_file $i
 done
 
-git submodule sync
-git submodule init
+echo "updating submodules..."
+git submodule --quiet sync
+git submodule --quiet init
+git submodule --quiet foreach git pull -q origin master
 git submodule update --merge --recursive
